Blood particles, particle system sort order, blood tornado

+++ RTS/stdafx.cpp
@@ -3,7 +3,7 @@
 #include <SDL.h>
 
 UNIT_SPACE(SECONDS) f64 sTotalTimeSeconds; ///< Total time since the update/draw loop started.
-UNIT_SPACE(SECONDS) f64 sElapsedSecondsSinceLastFrame; ///< Elapsed time of the previous frame.
+UNIT_SPACE(SECONDS) f32 sElapsedSecondsSinceLastFrame; ///< Elapsed time of the previous frame.
 
 bool s_debugToggle = false;
 bool s_wasTogglePressed = false;
